프로젝트/직원 수정 작업 완료 URL 접근 방지 뒤로가기 막기 팝업창 크기 조절 팝업창 수정 후 창 닫기
나머지 알아서 해라

+++ src/main/java/org/leedh/project/controller/PjtController.java
@@ -2,7 +2,6 @@ package org.leedh.project.controller;
 
 import lombok.extern.slf4j.Slf4j;
 import org.leedh.project.service.PjtService;
-import org.leedh.project.vo.PjtCommonVO;
 import org.leedh.project.vo.PjtVO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -12,6 +11,7 @@ import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @Controller
@@ -26,15 +26,39 @@ public class PjtController {
         this.pjtService = pjtService;
     }
 
+    // 프로젝트 홈 화면
+    @RequestMapping(value = "/main", method = RequestMethod.GET)
+    public String main(HttpServletRequest request, HttpSession session) {
+
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
+        return "/project/main";
+    }
+
     // 프로젝트 등록 페이지
     @RequestMapping(value = "/pjtRegister", method = RequestMethod.GET)
-    public String registerGET() throws Exception {
+    public String registerGET(HttpServletRequest request, HttpSession session) throws Exception {
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+
+            return "redirect:/";
+        }
         return "/project/pjtRegister";
     }
 
     // 프로젝트 등록 처리
     @RequestMapping(value = "/pjtRegister", method = RequestMethod.POST)
-    public String register(PjtVO pjtVO, RedirectAttributes rttr) throws Exception {
+    public String register(PjtVO pjtVO, RedirectAttributes rttr, HttpSession session, HttpServletRequest request) throws Exception {
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
         pjtService.register(pjtVO);
         rttr.addFlashAttribute("Message", "PJT register success");
         return "redirect:/main";
@@ -42,54 +66,57 @@ public class PjtController {
 
     //프로젝트 조회(전체)
     @RequestMapping(value = "/pjtShow", method = RequestMethod.GET)
-    public String pjt(Model model) throws Exception {
+    public String pjt(HttpServletRequest request, HttpSession session, Model model) throws Exception {
+
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
 
         List<PjtVO> pjtVo = pjtService.pjtShow();
         model.addAttribute("pjtList", pjtVo);
         return "/project/pjtShow";
     }
-    
+
     //프로젝트 수정
     @RequestMapping(value = "/pjtEdit", method = RequestMethod.GET)
-    public String pjtEditGET() throws Exception{
-    	
-    	//List<PjtVO> pjtVo = pjtService.pjtShow();
-    	//model.addAttribute("pjtList",pjtVo);
-    	return "/project/pjtEdit";
+    public String pjtEditGET(HttpServletRequest request, HttpSession session) throws Exception {
+
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
+        return "/project/pjtEdit";
     }
-    
+
+    //프로젝트 수정
     @RequestMapping(value = "/pjtEdit", method = RequestMethod.POST)
-    public String pjtEdit(PjtVO pjtVO,Model model) throws Exception{
-    	
-    	pjtService.pjtEdit(pjtVO);
-    	model.addAttribute("pjtEdit",pjtVO);
-    	return "/project/pjtEdit";
+    public String pjtEdit(PjtVO pjtVO, Model model, HttpServletRequest request, HttpSession session) throws Exception {
+
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
+        pjtService.pjtEdit(pjtVO);
+        model.addAttribute("pjtEdit", pjtVO);
+        return "/project/pjtEdit";
     }
-    
-    /*// 프로젝트 관리 페이지
-    @RequestMapping(value = "/login", method = RequestMethod.GET)
-    public String login(@ModelAttribute("loginDTO") LoginDTO loginDTO) throws Exception {
-        return "/user/login";
-    }*/
-
-    //  model.addAttribute("PJT", pjtVO);
-
-
-//    @RequestMapping(value="/delete", method=RequestMethod.POST)
-//    public String delete(@RequestParam("delList") Integer[] delList) {
-//
-//        for (Integer no : delList) {
-//            log.debug(Arrays.toString(delList));
-//            pjtService.deletePjt(no);
-//            pjtService.deletePjtCode(no);
-//        }
-//        return "redirect:/project/pjtShow";
-//    }
-
 
     // 프로젝트 삭제
     @RequestMapping(value = "/pjtDelete", method = RequestMethod.GET)
-    public String delete(String no) throws Exception {
+    public String delete(String no, HttpSession session, HttpServletRequest request) throws Exception {
+
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
         pjtService.deletePjt(no);
         pjtService.deletePjtCode(no);
         return "redirect:/project/pjtShow";
@@ -97,7 +124,13 @@ public class PjtController {
 
     // 프로젝트 선택삭제
     @RequestMapping(value = "/pjtDelete")
-    public String checkDelete(HttpServletRequest request) throws Exception {
+    public String checkDelete(HttpServletRequest request, HttpSession session) throws Exception {
+
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
 
         String[] ajaxMsg = request.getParameterValues("valueArr");
 
@@ -110,10 +143,16 @@ public class PjtController {
 
     //프로젝트 조회(전체)
     @RequestMapping(value = "/pjtUnion", method = RequestMethod.GET)
-    public String pjtUnion(Model model) throws Exception {
+    public String pjtUnion(Model model, HttpSession session, HttpServletRequest request) throws Exception {
 
-        List<PjtVO> pjtVo = pjtService.memberunion();
-        model.addAttribute("memberunion", pjtVo);
+        // url 직접 접근시 로그인 화면으로 리다이렉션 /  session.invalidate()로 세션 소거
+        if (request.getHeader("REFERER") == null) {
+            session.invalidate();
+            return "redirect:/";
+        }
+
+        List<PjtVO> pjtVo = pjtService.memberUnion();
+        model.addAttribute("memberUnion", pjtVo);
         return "/project/pjtUnion";
     }
